Use Arc instead of Cow
Cow is just copy when you have an owned type, pfff

We *could* use Rc instead of Arc in this case, all the Strings get freed
at the same time, and only once, but...

Rust complains, and technically it's true that they *could* be freed in
different threads, which *could* lead to a race condition, which could
be bad. It's probably not that much of a performance hit anyways, so ehh.